refactor(racingcar) : 비지니스 로직을 가진 객체는 domain 패키지, UI관련 객체는 view 패키지로 분리
