+++ src/main/kotlin/racingcar/RacingCarConsoleApplication.kt
@@ -1,5 +1,10 @@
 package racingcar
 
+import racingcar.domain.RaceHost
+import racingcar.view.RacingCarInputView
+import racingcar.view.RacingCarOutputView
+import racingcar.view.RacingCarRequest
+
 fun main() {
     val racingCarRequest: RacingCarRequest = RacingCarInputView.input()
     val racingResults = RaceHost.play(
