+++ src/main/kotlin/racingcar/TODO.md
@@ -40,4 +40,8 @@
   * [x] 더 이상 참가자 수를 입력받지 않는다.
   * [x] 자동차 이름들을 입력받을 때는 쉼표(',')를 기준으로 구분한다. ex) pobi, jason
   * [x] 자동차 이름들의 입력 값은 빈값이거나 공백일 수 없다.
-* [ ] 자동차 경주 게임이 끝나고 누가 우승했는지를 알려준다. (우승자는 한 명 이상일 수 있다.즉, 공동 우승이 가능하다.)
+* [x] 자동차 경주 게임이 끝나고 누가 우승했는지를 알려준다. (우승자는 한 명 이상일 수 있다.즉, 공동 우승이 가능하다.)
+
+## 자동차 경주 패키지 리팩터링
+* [x] 비지니스 로직을 가진 객체는 domain 패키지에 위치해야한다.
+* [x] UI와 관련된 객체는 view 패키지에 위치해야한다.
