+++ src/main/kotlin/racingcar/domain/RaceHost.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 object RaceHost {
